(a) give Constants and Util methods their own home in a shared modules folder (b) remove hard-coded 'data_pipeline' from everywhere
