+++ modules/RelationalDataLoader.py
@@ -2,7 +2,7 @@ import logging
 import argparse
 from sqlalchemy import create_engine
 from modules.DataLoadManager import DataLoadManager
-from modules.Shared import Constants
+from modules.shared import Constants
 from modules.data_load_tracking.DataLoadTrackerRepository import DataLoadTrackerRepository
 from modules.data_sources.DataSourceFactory import DataSourceFactory
 from sqlalchemy.orm import sessionmaker
