+++ modules/data_load_tracking/DataLoadExecution.py
@@ -2,13 +2,14 @@ from sqlalchemy import Column, DateTime, Integer, String, Boolean, BigInteger
 from sqlalchemy.sql import func
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.dialects.postgresql import UUID
+from modules.shared import Constants
 
 Base = declarative_base()
 
 
 class DataLoadExecution(Base):
     __tablename__ = 'data_load_execution'
-    __table_args__ = {'schema': 'data_pipeline'}
+    __table_args__ = {'schema': '{Constants.DATA_PIPELINE_EXECUTION_SCHEMA_NAME}'}
     id = Column(Integer, primary_key=True)
     correlation_id = Column(UUID(as_uuid=True), nullable=True)
     model_name = Column(String(250), nullable=False)
