+++ modules/shared/Utils.py
@@ -1,0 +1,8 @@
+import importlib
+
+
+def create_type_instance(type_name):
+    module = importlib.import_module(type_name)
+    class_ = getattr(module, type_name)
+    instance = class_()
+    return instance
